Normal

+++ app/src/main/java/com/adrapps/mytasks/api_calls/DeleteTask.java
@@ -1,9 +1,12 @@
 package com.adrapps.mytasks.api_calls;
 
+import android.Manifest;
+import android.content.Context;
 import android.os.AsyncTask;
 
-import com.adrapps.mytasks.presenter.TaskListPresenter;
 import com.adrapps.mytasks.R;
+import com.adrapps.mytasks.domain.Co;
+import com.adrapps.mytasks.presenter.TaskListPresenter;
 import com.google.api.client.extensions.android.http.AndroidHttp;
 import com.google.api.client.googleapis.extensions.android.gms.auth.GoogleAccountCredential;
 import com.google.api.client.googleapis.extensions.android.gms.auth.GooglePlayServicesAvailabilityIOException;
@@ -14,15 +17,19 @@ import com.google.api.client.json.jackson2.JacksonFactory;
 
 import java.io.IOException;
 
+import pub.devrel.easypermissions.EasyPermissions;
+
 public class DeleteTask extends AsyncTask<String, Void, Void> {
 
     private com.google.api.services.tasks.Tasks mService = null;
     private Exception mLastError = null;
     private TaskListPresenter mPresenter;
     private String listId;
+    Context context;
 
-    public DeleteTask(TaskListPresenter presenter, GoogleAccountCredential credential,
+    public DeleteTask(Context context, TaskListPresenter presenter, GoogleAccountCredential credential,
                       String listId) {
+        this.context = context;
         this.mPresenter = presenter;
         this.listId = listId;
         HttpTransport transport = AndroidHttp.newCompatibleTransport();
@@ -46,6 +53,7 @@ public class DeleteTask extends AsyncTask<String, Void, Void> {
         }
         return null;
     }
+
     @Override
     protected void onProgressUpdate(Void... values) {
         super.onProgressUpdate(values);
@@ -87,7 +95,14 @@ public class DeleteTask extends AsyncTask<String, Void, Void> {
 
 
     private void removeTask(String taskId) throws IOException {
-            mService.tasks().delete(listId,taskId).execute();
+        if (EasyPermissions.hasPermissions(context, Manifest.permission.GET_ACCOUNTS)) {
+            mService.tasks().delete(listId, taskId).execute();
             mPresenter.deleteTaskFromDatabase(mPresenter.getIntIdByTaskId(taskId));
+        } else {
+            EasyPermissions.requestPermissions(
+                    context, context.getString(R.string.contacts_permissions_rationale),
+                    Co.REQUEST_PERMISSION_GET_ACCOUNTS,
+                    Manifest.permission.GET_ACCOUNTS);
+        }
     }
 }
